+++ packages/style/src/helpers/index.ts
@@ -16,8 +16,6 @@ import isNestedSelector from './isNestedSelector';
 import isSupportsRule from './isSupportsRule';
 import isUnitlessProperty from './isUnitlessProperty';
 import isVariable from './isVariable';
-import processProperties from './processProperties';
-import processValue from './processValue';
 
 export {
   createAtomicCacheKey,
@@ -38,6 +36,4 @@ export {
   isSupportsRule,
   isUnitlessProperty,
   isVariable,
-  processProperties,
-  processValue,
 };
