+++ packages/style/src/next/client.ts
@@ -1,0 +1,34 @@
+/* eslint-disable sort-keys */
+
+import { isSSR } from '@aesthetic/utils';
+import { SheetType, StyleRule, SheetMap } from '../types';
+
+function getStyleElement(type: SheetType): StyleRule {
+  // This is a little hacky, but hopefully this never gets interacted with
+  // istanbul ignore next
+  let element = (isSSR()
+    ? { sheet: { cssRules: [], insertRule() {} } }
+    : document.getElementById(`aesthetic-${type}`)) as HTMLStyleElement;
+
+  if (!element) {
+    const style = document.createElement('style');
+    style.setAttribute('id', `aesthetic-${type}`);
+    style.setAttribute('type', 'text/css');
+    style.setAttribute('media', 'screen');
+    style.setAttribute('data-aesthetic-type', type);
+
+    document.head.append(style);
+    element = style;
+  }
+
+  return ((element as HTMLStyleElement).sheet as unknown) as StyleRule;
+}
+
+export function createStyleElements(): SheetMap {
+  return {
+    // Order is important here!
+    global: getStyleElement('global'),
+    standard: getStyleElement('standard'),
+    conditions: getStyleElement('conditions'),
+  };
+}
